Refactor JS-specific strategy code into a module per @joshrowley's suggestion in https://github.com/learn-co/learn-test/pull/42#discussion_r129170683
