+++ lib/learn_test.rb
@@ -27,6 +27,7 @@ require_relative 'learn_test/dependencies/firefox'
 require_relative 'learn_test/dependencies/green_onion'
 
 require_relative 'learn_test/strategy'
+require_relative 'learn_test/js_strategy'
 require_relative 'learn_test/strategies/jasmine'
 require_relative 'learn_test/strategies/python_unittest'
 require_relative 'learn_test/strategies/rspec'
